refactor: reduce more redundant code with new func created in utils.ts and move all types to types.ts
