+++ packages/module-manager/lib/runModuleManager.ts
@@ -5,6 +5,7 @@ import { select } from '@inquirer/prompts';
 import { execSync } from 'node:child_process';
 import { resolve } from 'node:path';
 import { readFileSync, writeFileSync } from 'node:fs';
+import type { ActionType } from './types.js';
 
 const manifestPath = resolve(import.meta.dirname, '..', '..', '..', 'chrome-extension', 'manifest.ts');
 
@@ -12,7 +13,7 @@ const manifestObject = JSON.parse(JSON.stringify(manifest)) as chrome.runtime.Ma
 const manifestString = readFileSync(manifestPath, 'utf-8');
 
 const runModuleManager = async () => {
-  const tool = await select({
+  const tool: ActionType = await select({
     message: 'Choose a tool',
     choices: [
       { name: 'Delete Feature', value: 'delete' },
